Temporarily reverting sofia-sip library back to prior to the changes merged on 2/7/2007 due to ongoing unresolved issues in the new code.  Will re-commit these changes once we receive fixes from the sofia-sip tree for these outstanding issues.  Make sure recommended.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@4308 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/sofia-sip/libsofia-sip-ua/nta/test_nta.c
@@ -654,7 +654,7 @@ int readfile(FILE *f, void **contents)
 
   *contents = buffer;
 
-  return (int)len;
+  return len;
 }
 
 #if HAVE_DIRENT_H
